Moving comparable.py into brewtils.test
