+++ test/schema_parser_test.py
@@ -12,11 +12,11 @@ from pytest_lazyfixture import lazy_fixture
 
 from brewtils.models import System
 from brewtils.schema_parser import SchemaParser, BrewmasterSchemaParser
-from test.utils.comparable import (
+from brewtils.test.comparable import (
     assert_parameter_equal, assert_command_equal, assert_system_equal,
     assert_instance_equal, assert_request_equal, assert_patch_equal,
     assert_logging_config_equal, assert_event_equal, assert_queue_equal,
-    assert_principal_equal, assert_role_equal, assert_job_equal
+    assert_principal_equal, assert_role_equal, assert_job_equal,
 )
